backend/sovoltosegunda/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
decorator==5.1.1
Django==4.0.3
djangorestframework==3.13.1
docutils==0.15.2
executing==0.8.3
flake8==4.0.1
gunicorn==20.1.0
ipython==8.2.0
jedi==0.18.1
jmespath==0.10.0
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
